Create Index.html
Skillsoft issued completion badges are earned based on viewing the percentage required or receiving a passing score when assessment is required. Software testing is a crucial part of software development. Examine testing objectives, principles and processes, the psychology and ethics of testing, and different development models in a software life cycle.